updated FlowscaelFlowUpdate

+++ backend/edu.iu.incntre.flowscaleflowupdate/src/main/java/edu/iu/incntre/flowscaleflowupdate/LoadFlow.java
@@ -47,7 +47,7 @@ public class LoadFlow {
 		this.loadedPort = loadedPort;
 	}
 	
-	public void setPacketCount(double packetCount){
+	public void setPacketCount(long packetCount){
 		
 		this.packetCount = packetCount;
